refactor: ルーティングパッケージの命名を他のパッケージに合わせて単数形に変更
