+++ internal/router/health_check.go
@@ -1,4 +1,4 @@
-package routes
+package router
 
 import (
 	"net/http"
