+++ internal/router/openapi.go
@@ -1,4 +1,4 @@
-package routes
+package router
 
 import (
 	"github.com/gin-gonic/gin"
